Part 1 of quest refactor
