+++ bundles/ranvier-quests/quest-goals/BountyGoal.js
