+++ bundles/ranvier-quests/quest-goals/EquipGoal.js
