+++ bundles/ranvier-quests/quest-goals/KillGoal.js
