+++ src/QuestGoalRegistry.js
@@ -1,0 +1,6 @@
+/**
+ * Simple map of quest goal name => class definition
+ */
+class QuestGoalRegistry extends Map {}
+
+module.exports = QuestGoalRegistry;
